ml_worker/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.31.1
pillow==10.4.0
transformers==4.45.2
torch==2.4.1
python-multipart==0.0.12
opencv-python-headless==4.10.0.84
imagehash==4.3.1
